requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Caching
redis==5.0.1
django-redis==5.4.0

# Environment variables
python-decouple==3.8

# HTTP requests for TMDb API
requests==2.31.0

# Development
python-dotenv==1.0.0